Fixed a but with the correlation of clusters. The problem was there since the beginning and never discovered because we have always tested telescope setup with detector of the same kind.
After a test with a mixed setup (5 MimoTel and 1 Mimosa18), it becames clear
that during the booking and the filling of the histograms the meaning of
the X and Y axes were exchanged.

Even if this bug was present since the beginning, it didn't affect any
calculation or any result.