Finally a version of pysub The EUTelescope Python submitter that I like. Completely refurbished the previous version because it was too messy to handle. Now it is object oriented, well defined into a package and using as many python libraries as possible.
I'm starting moving from the old test/pysub folder all the submitters to this folder
because this is the right place. I'm also trying to have them documented with the
same Doxygen file. This would be beautiful.

Interesting new features:
--> Added optparse to simplify arguments and options
--> Added ConfigParse to simplify the handling of different configuration
--> Added logging to have everything logged into a rotate file
--> All the commands (but the GRID ones) are taken from libraries and not
    using external commands

While porting the old submitters to the style I'll also add full GRID submission.

This has to be considered still a work in progress commit for the submit-converter.py