adding fittrackvec and fithitvec collections from the daughter classes here. clear and get methods added. Both EUTelKalmanFilter and EUTelGBLFitter should be using these vectors.
+        // contains the fitted tracks, accessible through class methods
+        IMPL::LCCollectionVec * _LCIO_fittrackvec;
+
+        // contains the fitted hits, accessible through class methods
+        IMPL::LCCollectionVec * _LCIO_fithitvec;

_LCIO_fithitvec seems to be redundant.
the current plan is to have EUTelTrackImpl with a collection of EUTelTrackStateImpl as out definition of a track.
EUTelTrackStateImpl contains already track fit information as a track state at the scattering volume surface,
independently in the EUTelTrackImpl object there is a list of references to measured hit objects. which should be actually placed in EUTelTrackStateImpl objects.