GBLfit: Add DUT histograms for residuals
Used same scheme as for telescope layers in declaring histograms.
Currently histograms are filled according to the state location, which is
equal to the plane identifier (some number for a DUT). This is hard coded in
the GBL histogramming step, otherwise no histogram is filled. This could be
changed in the future to fill histograms for all existing residual entries in
the map.