Updated fitter classes to hgjersdal branch, and did the minimal changes to the official processors for them to compile.
The fitter code, implemented in EUTelDafTrackerSystem and EUTelDafEigen, are now template classes. The first argument it the type of the data in the fitter, should be float or double. Using floats greatly improves speed. The Kalman filter is not very numericallt stable, during development shiching to doubles can help if any problems arise. The second argument is the dimension of the fitter. It currently only works with 4 parameters, but it should be possible to implement a 5D fitter for magnetic fields while reusing much of the code.

The function dumpToAscii was removed.