Bugsandcosmetics #299 -- Squashed commits contained in pull request #299 https://github.com/eutelescope/eutelescope/pull/299 -- Description:
Point and getImpactPoint take care of that. The new function
getImpactPoint has two advantages. First, it allows for an easy usage of
a track slope. So far, the track direction is set to be in
z-direction (i.e. zero slope), but in principle one could add the slope
of a track here. Secondly, the new function does not require matrix
inversion, which makes it more efficient.

EUTelGeometryTelscopeGeoDescription.cpp
2.) In order to make the rotation-order consistent, the
EUTelGeometryTelscopeGeoDescription.cpp was changed in order to have the
rotation-order: first around the global X-axis, then around the global
Y-axis and finally around the global Z-axis. Since this also goes into
the HitMaker, all rotations in the HitMaker now have the order
X,Y,Z. After this change, the rotation order of the HitMaker, the
EUTelGeometryTelscopeGeoDescription, the Alignment (EUTelMille.cc), the
EUTelProcessorApplyAlignment.cc and the EUTelTestFitter.cc are the
same. If this change should be rejected, the rotation order in all
mentioned files should be checked for consistency. Additionally, the
rotation order in our next pull request, from the branch analysisupdate,
also has to be changed correspondingly. I also suggest to document
somewhere, which rotation-order is used. On long-term basis one could
also consider making the rotation order a parameter, for example from
the config-file.

EUTelMille.cc
3.) The binsize and the plot ranges of the histograms drawn by
EUTelMille.cc were increased, since reading the resulting histograms
otherwise was difficult for our data.

steering-template files
4.) Some purely cosmetic changes were applied to some steering-template
files.